+++ test/config/ODM/application.config.php
@@ -10,7 +10,6 @@ return [
     'modules' => [
         'DoctrineModule',
         'DoctrineMongoODMModule',
-        'Phpro\DoctrineHydrationModule',
         'Laminas\ApiTools',
         'Laminas\ApiTools\Admin',
         'Laminas\ApiTools\Hal',